Run visual CLI from main
