+++ main.py
@@ -1,0 +1,7 @@
+"""Application entry point that launches the visual command line interface."""
+
+from visualization_osmnx import run_visual_cli
+
+
+if __name__ == "__main__":
+    run_visual_cli()
